feat(dashboard): InspectionPage `Suspense` 적용 (#202)
